fixed stopAt in ExploreRecursive. Added tests
